feat: create a general crypto fiat input
